move the telemetry hanging test to an integration test
